feat: support server-stream
gRPC supports four kinds of RPCs: unary, client-stream, server-stream
and bidirectional. We want to support them all too.

This change adds support for server-stream RPCs (one message from the
client, a stream of messages from the server). This is pending real
support for knowing what the type of each function is, but should
unblock progress for now.