feat: Complete all RPC kinds (#28)
We have three of the four gRPC RPC kinds (rpc, stream, subscription).

This change adds the final RPC kind: upload (client-stream, single
message response). It also adds the option of adding an `init` type
to an RPC, where we can send a message of another schema before the main
stream, which is a very common pattern used in filesystem (i.e. open
file for writing, where the first message is the path and the other
messages are the chunks) and collaboration (i.e. first message is the
channel name and the rest of the bidirectional messages are the actual
stream).