Lockdown typebox peer dependency to patch
Typebox is a peer dependency because we want people to author their own types. Typebox is v0.x.x which means it can break at any change.

https://github.com/sinclairzx81/typebox/pull/689 0.32.0 seems to be a breaking change.

I chose to just lock down minor changes since that seems to match what typebox maintainer abides by.

I also added it as a dev dependency since you need it for development.

If locking down the version becomes an issue we can make typebox a direct dependency and re-export it for consumers (but that kinda sucks)