[river] Prevent drift between sessions and connections (#179)
We have a few invariant violations caused because of the fact that
during the initial connection establishment, we do not associate the
connection with the session. That means that anything that tries to
interact with the session and expects the connection to be there  (e.g.
the grace period timeout) will not see the connection there and proceed
to close the session but not the connection!

This change adds a new field to the session: `handshakingConn`, which is
used to hold the associated handshaking connection but without promoting
it to a full connection (which would cause some mayhem with the rest of
the state machinery). It still allows the transport to request for it to
be closed safely so that even if we are extremely unlucky and the grace
period elapses mid-handshake in another session, we can be safe about
the global state and won't leave the peer with a dangling connection
that cannot be used. Finally, it also ensures to clear any existing
grace period callbacks when installing a new one, to avoid the case
where a partial connection overwrites an existing grace period timeout,
and then the next time the client connects, the forgotten grace period
fires and causes mayhem.