Second attempt to clean up after reconnects
We had a previous attempt to clean up after reconnects. But there's
still one more failure mode where _something_ happens during the
reconnect handshake that makes the server believe it has never seen the
client before in its life and assign a new session id. When that
happens, the previous connection is never reset, and the timer
eventually fires and kills the new connection accidentally.

This change now makes the establishment of a new connection always clean
up the previous data in the session, no matter what.