Prevent session id mismatches upon reconnects
Currently if a client reconnects and finds that the server was restarted
without its knowledge, both sides just paper over this fact and carry
on. This causes trouble because now the client has yeeted the previous
session and internally replaced it with a new, blank session. That's all
good, but the client never told the server that this happened, which is
bad, because from now on, every time the client reconnects, they will
never agree on what the peer's session id should be, and it will cause
no transparent reconnects to be able to happen from then on. This
probably also causes some invariant violations, too.

This change now ensures that when the client detects a mismatch, it
proactively removes the session and reconnects the socket.