Avoid races when setting the grace period
We currently have a race where the `onDisconnect` callback can be
invoked arbitrarily late. This means that when replacing a new
connection, we can legally get the `onDisconnect` callback invoked after
we had already established the new connection, causing a) the grace
period to be installed after the reconnection was set and b) the grace
period could be invoked with a connection already in place!

This change tries _even harder_ to avoid races with the session <->
connection mapping. Now whenever `onDisconnect` is called with the wrong
parameters (either a connection already removed, or a mismatched
connection), or the grace period elapses with a connection already
installed, we just bail out and avoid making more damage.