Make the server reject misbehaving connections (#212)
## Why

We currently have a setup where the server just trusts whatever the
client requests, and stomps over whatever was already there. That's okay
in some cases, but not all! For example, if the client's timers are all
wonky (like they are when the tab is hidden), we might get in a
situation where the client tries to reconnect after its server-side
grace period has been exceeded, and they don't agree on what the state
should be. Also, if pid2 restarts for whatever reason, both parties
should quickly agree that they're not agreeing upfront.

## What changed

This change now adds a new (optional) field to the handshake,
`expectedState`. The client should fill this up and explicitly tell the
server whether this is a reconnect attempt (currently this is done by
checking if the `advertisedSessionId` is set, but that can be done by
explicitly tracking whether the session has seen the server at least
once when we move away from this model), and tell the server what its
expected next sequence number is (so that the server can also detect
drifts in this). If the clients don't fill this information up, the
server will behave in a backwards-compatible fashion for safety.

## Versioning

- [ ] Breaking protocol change :massage_man: (backwards-compatible)
- [ ] Breaking ts/js API change :massage_man: