Make the `nextExpectedSeq` test stricter
We used to have a loose bound of obviously wrong `nextExpectedSeq`s. But
we could still get this check wrong if we got unlucky and didn't have
the right entry in the buffer.

This change now checks that the `nextExpectedSeq` is either in the
`sendBuffer` ready to be retransmitted _or_ exactly matches the server's
next `seq`.