Don't "fail to write" when buffering
Currently the `send` implementation for UDS "fails to write" (i.e.
`send` returns `false`) if there is any buffering done. But that's not
the intention of the code!

This change now makes the `send` implementation of UDS consistent with
the WS one: it always returns `true` _unless_ we can reliably tell that
the connection was closed a priori.