Use a state machine to manage session state transitions (#220)
## Why

Keeping track of what handlers are registered when and knowing when to
cleanup/register new callbacks and fields in a session's lifecycle is
really difficult. Let's encode valid states via a state machine and make
illegal states unrepresentable.

This PR also addresses two other problems that are heavily tied to the
session lifecycle:
- Message order synchronization during handshake: we used to synchronize
`nextExpectedSeq` in the handshake which is useful to detect desyncs in
server to client direction (that is, if the server was about to send a
message that would cause the client to receive an out-of-order message,
we should just close the session there instead of continuing). However,
we never did this check in the other direction.
- Relying on browser timers: we used to rely on a client-side
`setInterval` to send heartbeat messages to the server and the server
would kill the client if it didn't send heartbeats frequently enough.
Turns out that Chromium _heavily_ throttles intervals when backgrounded
(source:
https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/renderer/platform/scheduler/main_thread/page_scheduler_impl.h;l=54?q=kDefaultThrottledWakeUpInterval&ss=chromium%2Fchromium%2Fsrc)
so a `setInterval` for, say 3s, could take up to 60s to actually fire.
This meant connections were getting killed far too frequently. However,
if we make the timers server authoritative, the client can always
respond to server heartbeats as data-listener-based wakeups are not
throttled :)

## What changed

- Implement a state machine that supports async-transitions via a
callback-based system. We use a proxy to ensure that session states
we've transitioned from cannot be used again.
- Each state registers what to do on state exit and on close (exiting
from the state machine entirely, e.g. closing the session object) so we
have consistent cleanups that don't leak. This has been the source of
quite a few bugs.
- Made heartbeats server authoritative.
- Clients no longer schedule their own heartbeats. Instead, it responds
directly to any server-sent heartbeats.
- Servers are responsible for closing a connection on too many heartbeat
misses.
- This is ok even in the case of phantom disconnects as closing the
underlying TCP conn will still propagate in the case of a WebSocket
transport (and UDS cannot have phantom disconnects)
- Update the handshake messages
- No longer sends `reconnect`, this can be inferred a combination of
`nextExpectedSeq` and `nextSentSeq`
- Now relies on `sessionId` to generate the ID on the session on the
server side so the IDs of the client session and the server session
match
- Now sends `nextSentSeq`. We do some more comparison here to check for
mismatched session state before calling the handshake ok. Specifically,
there are two cases where we cannot accept a session:
	  - the client is in the future (`client.nextSentSeq > server.ack`)
	  - server is in the future (`server.seq > client.nextExpectedSeq`)
- Distinguish retriable handshake failures from fatal ones.
- Only a session state mismatch is retriable (i.e. kill the session and
start over), everything else should be 'fatal' and stop retrying.
- Testing infra changes
- Made constructing a transport not automatically connect by default. We
had cases where we registered listeners _after_ connecting and would
miss some events.
- No longer expose connections to the transport (it doesn't manage
these! it shouldn't be able to introspect connections). I pulled these
out into test helpers `numberOfConnections(transport)` and
`closeAllConnections(transport)` which covers our existing use cases.

## Versioning

- [ ] Breaking protocol change
- [x] Breaking ts/js API change

<!-- Kind reminder to add tests and updated documentation if needed -->