Allow the custom handler to return more error codes
Previously the only allowed error in a custom handler is `false` (which
doesn't say a lot). But we want to be able to return something more
expressive.

This change now adds a new type for this, and allows the custom handler
to return more responses. This allows us to have different behavior for
different rejection kinds (some more permanent than others).