Allow the custom handler to return more error codes (#238)
## Why

Previously the only allowed error in a custom handler is `false` (which
doesn't say a lot). But we want to be able to return something more
expressive.

## What changed

This change now adds a new type for this, and allows the custom handler
to return more responses. This allows us to have different behavior for
different rejection kinds (some more permanent than others).

## Versioning

- [ ] Breaking protocol change
- [X] Breaking ts/js API change (return type from the validator changed)