Make treatment of session disconnect and transport closing similar (#323)
## Why

After the transport closes, calling a procedure from the client leads to
throwing an error telling you transport is closed. This happens because
we end up creating a new session and then failing at getting the bound
send

## What changed

We no longer expose the status of the transport to the caller, we
pretend that your session suddenly disconnected, unifying the interface
when reading after session closed and calling after transport closed.

## Versioning

- [ ] Breaking protocol change
- [x] Breaking ts/js API change....kinda

<!-- Kind reminder to add tests and updated documentation if needed -->