Refactor ServiceContext as generic (#322)
## Why

When we are creating a server using river, we have to declare the `ServiceContext` interface globally so TypeScript can infer the type for `ctx` within a procedure. 

```js
 declare module '@replit/river' {
   interface ServiceContext {
       db: Database;
   }
 }
```

However, it will be an issue in a mono repo setup where we want to create multiple servers with different `ServiceContext`. 

<!-- Describe what you are trying to accomplish with this pull request -->

## What changed

Use builder pattern for `ServiceSchema` so we can create a `ServiceSchema` with Context like 

`const ServiceSchema = createServiceSchema<Context>();`

Example: 

```js

const extendedContext = {
  testctx: Math.random().toString(),
};

const ServiceSchema = createServiceSchema<typeof extendedContext>();

const services = {
  testservice: ServiceSchema.define({
    testrpc: Procedure.rpc({
      requestInit: Type.Object({}),
      responseData: Type.String(),
      handler: async ({ ctx }) => {
        return Ok(ctx.testctx);
      },
    }),
  }),
};
```

When we create a server, it will perform type checking to make sure that all the services are expecting the same `ServiceContext` as the type of `extendedContext` in the config.

```js
createServer(serverTransport, services, {
  extendedContext,
});
```

There is no type-checking for `Context` in createClient(). We are using `any` to skip it behind the scene because it is a implementation detail on the server side.

```js
const client = createClient<typeof services>(
  clientTransport,
  serverTransport.clientId,
); 
```

<!-- Describe the changes you made in this pull request or pointers for the reviewer -->

## Versioning

- [ ] Breaking protocol change
- [x] Breaking ts/js API change

<!-- Kind reminder to add tests and updated documentation if needed -->