Update Github Workflow (#327)
## Why

The previous `release drafter` and `Build and upload package` workflow
did not work! In the previous setup, when we publish a draft release
created by `release drafter`, it does not update the version in
`package.json` so the `package.json` file in the released source code is
outdated.

## What changed

- We still use release drafter like before. But instead of publishing
the draft release directly, we should Go to Actions ->
workflows/prepare-release.yml and run the workflow.
- It will automatically create a git tag from the draft release 
- The new git tag will trigger the `publish.yml` workflow, which will
build and publish the package to npm
- Use node 22 in CI

## Versioning

- [ ] Breaking protocol change
- [ ] Breaking ts/js API change

<!-- Kind reminder to add tests and updated documentation if needed -->

---------

Co-authored-by: graphite-app[bot] <96075541+graphite-app[bot]@users.noreply.github.com>